+++ CHANGES.md
@@ -19,6 +19,8 @@ v1.6.2 - YYYY-MM-DD
 - Added missing range checks to `pdfioArrayCopy` and `pdfioDictCopy`.
 - Refactored PDF encryption code to fix unlocking with certain files.
 - Improved xref table loop detection (Issue #148)
+- Fixed xref reconstruction for objects lacking a `Type` value.
+- Fixed `pdfioPageOpenStream` for indirect `Contents` arrays.
 - Fixed an error propagation bug when reading too-long values (Issue #146)
 - Fixed a Clang warning.